requirements.txt
Flask==2.0.2
requests==2.26.0
ptvsd==4.3.2 # Required for debugging.
google-cloud-bigquery==1.28.0
google-cloud-monitoring==2.8.0
protobuf3==0.2.1